FIX: duplicati - allow restore of docker persistent data
- adds ability to restore to docker_home when duplicati_data_permissions is overridden to "rw"

NOTE: I'm not sure why this wasn't addressed in https://github.com/davestephens/ansible-nas/pull/204. I addressed it with a question in https://github.com/davestephens/ansible-nas/issues/171 but never received a response. Hopefully this is a desirable enhancement allowing direct restoring of the container persistent data without having to restore to a temp area and then move them.